Cogroo autostart (#1)
* Autoloading cogroo

* Changing CI script

* Implementing tests

* Implementing codecov report
