+++ .travis.yml
@@ -1,0 +1,14 @@
+language: java
+jdk:
+  - openjdk8
+python:
+  - 3.6.7
+before_install:
+  - pyenv shell 3.6.7
+install:
+  - pip install pytest-cov pytest codecov
+script:
+  - python setup.py install
+  - py.test --cov-report term --cov=cogroo_interface
+after_success:
+  - codecov
