+++ setup.cfg
@@ -1,0 +1,3 @@
+# Inside of setup.cfg
+[metadata]
+description-file = README.md
